tag is NSInteger* not int* in recent AppKit
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@39581 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSWorkspace.m
@@ -1054,7 +1054,7 @@ static NSDictionary		*urlPreferences = nil;
 		       source: (NSString*)source
 		  destination: (NSString*)destination
 		        files: (NSArray*)files
-			  tag: (int*)tag
+			  tag: (NSInteger*)tag
 {
   id app;
